refactor: climate tool folder structures
