+++ apps/picsa-tools/climate-tool/src/app/components/chart-tools/combined-probability/combined-probability.scss
