+++ apps/picsa-tools/climate-tool/src/app/components/chart-tools/crop-analysis/crop-analysis.html
