+++ apps/picsa-tools/climate-tool/src/app/components/chart-tools/line-tool/line-tool.component.html
