+++ apps/picsa-tools/climate-tool/src/app/components/chart-tools/probability-tool/probability-tool.scss
