+++ apps/picsa-tools/climate-tool/src/app/components/chart-tools/probability-tool/probability-tool.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnChanges } from '@angular/core';
-import { IProbabilities } from '../../models';
+import { IProbabilities } from '../../../models';
 
 @Component({
   selector: 'climate-probability-tool',
