+++ apps/picsa-tools/climate-tool/src/app/components/climate-tool-components.module.ts
@@ -9,14 +9,15 @@ import { PicsaTranslateModule } from '@picsa/shared/modules/translate';
 
 import { PicsaClimateMaterialModule } from './material.module';
 
+import { CombinedProbabilityComponent } from './chart-tools/combined-probability/combined-probability';
+import { CropAnalysisComponent } from './chart-tools/crop-analysis/crop-analysis';
+import { LineDatePickerHeaderComponent } from './chart-tools/line-tool/line-date-picker-header';
+import { LineToolComponent } from './chart-tools/line-tool/line-tool.component';
+import { ProbabilityToolComponent } from './chart-tools/probability-tool/probability-tool';
+import { TercilesToolComponent } from './chart-tools/terciles-tool/terciles-tool.component';
+
 import { ClimateChartLayoutComponent } from './chart-layout/chart-layout';
 import { ClimateChartOptionsComponent } from './climate-chart-options/climate-chart-options.component';
-import { CombinedProbabilityComponent } from './combined-probability/combined-probability';
-import { CropAnalysisComponent } from './crop-analysis/crop-analysis';
-import { LineToolComponent } from './line-tool/line-tool.component';
-import { LineDatePickerHeaderComponent } from './line-tool/line-date-picker-header';
-
-import { ProbabilityToolComponent } from './probability-tool/probability-tool';
 import { ClimateShareDialogComponent } from './share-dialog/share-dialog.component';
 import { ViewSelectComponent } from './view-select/view-select';
 import { ClimatePrintLayoutComponent } from './print-layout/print-layout.component';
@@ -31,6 +32,7 @@ const Components = [
   LineToolComponent,
   LineDatePickerHeaderComponent,
   ProbabilityToolComponent,
+  TercilesToolComponent,
   ViewSelectComponent,
 ];
